Cookie issue resolved, and routes updated.
